Add 5.1 source code in a tags file
In order to simplify the building of code inside of docker,
we store shallow tags of the main OME repositories in a single
file.  The created images can be used as data containers into
which one can merge other branches to build custom code. As
decoupling continues, each "tag" can become a representation
of an entire release, i.e. which tags of which repositories
have been tested together.